Improved Loot Drops
Fixed a problem where players were still able to create alliances
despite

    PreventTribeAlliances=True

being set. Apparently, this does not have the desired effect. I had
to add

    MaxAlliancesPerTribe=0
    MaxTribesPerAlliance=0

to disable creation of tribe alliances. I found out that the other
setting did not work when a player griefed another player by
creating an alliance, coercing the player to put stuff in storage,
and then kicking the player from the alliance.

Changed the Blue, Green, and Yellow Cave and Ice Cave loot drops
to provide items that might be useful to players who get their
armor busted or die in a cave. For example, Blue loot drops now
include a crossbow and grappling hooks to reach bodies in the lava
in Cave of the Massive.

Changed the Purple Supply Drops to provide imprinting food 50% of
the time and Tek Generators 50% of the time. Hopefully, the Tek
Generators will reduced the amount of wiring used in large bases
and prevent lag.

Changed the Red Supply Drops to provide one piece of low level Tek
Armor. Putting valuable items in Red Supply Drops caused people to
build bases that were used specifically to camp supply drops. One
player installed teleporters and we didn't see a Red Supply Drop
for three days.

Changed the Broodmother and Megapithecus Loot Drops to provide
three trophies so that multiple players can get a trophy. I like
to invite new players to boss fights so they can see how they work
and it is nice to be able to give them a trophy afterward.
Hopefully, this will give them an incentive to attempt the boss
fights on their own.